requirements/dev.txt

-r base.txt

ipython
pytest-django
factory_boy
